# src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FIAP Video Processor API",
    "version": "1.0.0",
    "description": "API para gerenciamento de vídeos enviados por usuários e notificação de upload"
  },
  "paths": {
...
    "/frames/{frameId}": {
      "get": {
        "summary": "Obtém um frame pelo ID",
        "tags": ["Frames"],
        "parameters": [
          {
            "name": "frameId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhes do frame retornados com sucesso"
          },
          "404": {
            "description": "Frame não encontrado"
          },
          "500": {
            "description": "Erro ao consultar o frame"
          }
        }
      }
    },
...
  }
}